pyproject.toml
[project]
name = "heart-disease-classification"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "joblib>=1.5.2",
    "kagglehub[pandas-datasets]>=0.3.13",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.6",
    "numpy>=2.2.6",
    "pandas>=2.3.2",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "xgboost>=3.0.5",
]
